Corrige la personnalisation du potentiel de points d'une action
